Move basic auth handler into exposition. Update docstring.
